Split Parakeet ownership source contracts
